refactoring interfaces package to correct the exposed namespace iot.agile.Class
