+++ agile-interfaces/src/main/java/iot/agile/DeviceManager.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package iot.agile.agile.interfaces;
+package iot.agile;
 
 import java.util.Map;
