+++ iot.agile.DeviceManager/src/main/java/iot/agile/devicemanager/examples/ConnectDevice.java
@@ -18,7 +18,7 @@ package iot.agile.devicemanager.examples;
 import org.freedesktop.dbus.DBusConnection;
 import org.freedesktop.dbus.exceptions.DBusException;
 
-import iot.agile.agile.interfaces.Device;
+import iot.agile.Device;
 
 /**
  * @author dagi
