+++ iot.agile.ProtocolManager/src/main/java/iot/agile/protocolmanager/example/BLEDiscovery.java
@@ -18,7 +18,7 @@ package iot.agile.protocolmanager.example;
 import org.freedesktop.DBus.Error.ServiceUnknown;
 import org.freedesktop.dbus.DBusConnection;
 
-import iot.agile.agile.interfaces.ProtocolManager;
+import iot.agile.ProtocolManager;
 
 /**
  *
